Introduced the notion of an object address, which identifies the server node on which an object resides. Services will need to return an object address, to which clients can then easily subscribe.
Unfortunately, my dream of returning objects directly from service requests is
thwarted by the complexities of cases where the server handling the request is
not the owner of the object in question. In such cases, one cannot avoid the
double round trip, as the first round trip provides the address (which is on
another server) and the client must subsequently connect to said server, send a
subscription request, and obtain the object, prior to communicating success to
the callback.

I may still look into optimizing the single-server situation, but since object
subscriptions are so infrequent, it may not be worth the additional complexity.