Revamped the way attributes and objects are streamed. Since those are not immutable value objects, it's not feasible to take the "call the constructor with data read from the stream" approach.
Instead we define readContents and writeContents for all attributes, which
enables us to define a generic readContents/writeContents for all NexusObject
instances based on their ability to enumerate their attributes (this does mean
that only the attributes are streamed, but I think that's sensible).

Also added the ability to read and write a Streamable implementing Class, which
turned out to be needed to stream Address instances.