Added Nexus.routed() and implementation.
This provides a means to route a slot event notification into the execution
context (i.e. thread) of a particular entity.

Most of the time this is not needed, because an entity will be listening to an
attribute of an object registered in the same context as the entity, so it is
known that the events are dispatched in the correct execution context. However,
in some situations, like when listening for session disconnection, the
originator of the event is some other thread (like the network thread in the
example case), and such events must be routed into an entity's execution
context before they can trigger the mutation of entity state, etc.