We need to use an even newer version of Jetty.
You might think that 7.4.3.v20110701 was not newer than 8.0.0.M3, but you would
be mistaken. Jetty is currently being developed in parallel against the Servlet
2.5 and Servlet 3.1 specifications in Jetty 7x and 8x respectively.

However the 7x branch is the main line of development, and periodically they
take that code and convert it all to Servlet 3.1 and create an 8x milestone
release. The latest 8x milestone can handle WebSocket protocol versions up to
6. Chrome currently uses protocol version 8, which the latest Jetty 7x groks.

Whee!