Rejigger processor testing to avoid reusing compiler components.
This seems to cause random bizarre errors. Said errors still happen sometimes,
which is probably because we rerun the tests in the same JVM. Alas, forking the
unit tests seems to be beyond the capabilities of SBT, or rather beyond my
tolerance for fucking with SBT's internals to make it happen.