Added NexusObject.onLost; wired up unsubscribe.
NexusObject.onLost is triggered if an object is cleared (unregistered) on the
server while it still has subscribers. The subscribers (on the client) are
notified that the object is no longer around and their subscription info is
automatically cleared.

Currently onLost is not emitted when a connection is lost because connection
lost handling is not yet wired up. TODO!