Fixed issue with top-level packages.
Also added a test to ensure that non-value fields are not included in a
streamer (on the way to which I discovered and fixed the top-level package
bug).

We discourage the use of non-value fields in streamables (since they're not
streamed), but maybe you need to do something sneaky and we want to be sure not
to choke on them.