Wired up unsubscribing. Switched to a "subscription handle" approach.
It's long been a source of bugs in Narya that you have to unsubscribe by
passing in the object that you got in the first place. This enables
situations where you request to subscribe to an object, but then decide that
you don't wnat it before the subscription request completes.

The "I'm not thinking about this case" solution, which is often implemented, is
just to say "gee, I have no object, I'll just skip the unsubscribe request",
but then the object comes in later and is never unsubscribed from. The proper
solution is a bunch of backbending to ensure that when the object finally comes
in, you turn around and send a request to unsubscribe.

In Nexus, we do that backbending for you, and you get a handle the instant you
request your subscription, and you can unsubscribe any time, before or after
your subscription is completed.