Test that auto-created ctor failure is passed to requester.
If you issue a request that causes an entity to be auto-created and that entity
throws an exception in its constructor, we want that exception to be passed
back to the requester. And it is.

I realize now that the entity binding will leak, which is not a huge deal
(since it's small and retains no references), but I'd rather clean it up. So
that's coming next.